third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-window-open.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" name="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    appHistory.onnavigate = t.step_func_done(() => {
      assert_unreached("onnavigate should not have fired in source window");
    });
    iframe.contentWindow.appHistory.onnavigate = t.step_func_done(e => {
      assert_true(e.cancelable);
      assert_true(e.canRespond);
      assert_false(e.userInitiated);
      assert_false(e.hashChange);
      assert_equals(e.formData, null);
      e.preventDefault();
    });

    window.open("#", "i");
  });
}, "window.open() fires navigate event in target window but not source");
</script>
